requirements.txt
amqp==5.3.1 ; 
asgiref==3.9.1 ; 
billiard==4.2.1 ; 
celery==5.5.3 ; 
click-didyoumean==0.3.1 ; 
click-plugins==1.1.1.2 ; 
click-repl==0.3.0 ; 
click==8.2.1 ; 
colorama==0.4.6 ;
django-csp==4.0 ; 
django-environ==0.12.0 ; 
django-taggit==6.1.0 ; 
django==5.2.6 ; 
djangorestframework==3.16.1 ; 
djangorestframework-simplejwt==5.4.0 ; 
kombu==5.5.4 ; 
packaging==25.0 ; 
prompt-toolkit==3.0.52 ; 
psycopg2-binary==2.9.10 ; 
python-dateutil==2.9.0.post0 ; 
redis==6.4.0 ; 
six==1.17.0 ; 
sqlparse==0.5.3 ; 
tzdata==2025.2 ; 
vine==5.1.0 ; 
wcwidth==0.2.13 ; 
yt-dlp ; 
python-telegram-bot==22.4 ;
requests==2.32.5 ;
gunicorn==23.0.0 ; 
